Reviewed project changes. - Remove example/ - move qtest under test dir - move unit/base/ - remove unused refl methods.h
