--- example/package.json
@@ -1,4 +1,0 @@
-{
-    "name": "jest.config.js",
-    "type": "module"
-}
