+++ test/unit/base/conv/function.cpp
@@ -1,8 +1,8 @@
-#include "base/test/test.h"
-
 #include "base/conv/function.h"
 
 #include <vector>
+
+#include "../../util/test.h"
 
 using namespace Conv;
