+++ test/unit/base/conv/numtostr.cpp
@@ -1,6 +1,6 @@
 #include "base/conv/numtostr.h"
 
-#include "base/test/test.h"
+#include "../../util/test.h"
 
 using namespace test;
 using namespace Conv;
