+++ test/unit/base/conv/parse.cpp
@@ -1,6 +1,6 @@
-#include "base/test/test.h"
-
 #include "base/conv/parse.h"
+
+#include "../../util/test.h"
 
 using namespace test;
 using namespace Conv;
