+++ test/unit/base/conv/strtonum.cpp
@@ -1,6 +1,6 @@
 #include "base/conv/strtonum.h"
 
-#include "base/test/test.h"
+#include "../../util/test.h"
 
 using namespace test;
 using namespace Conv;
