+++ test/unit/base/conv/tostring.cpp
@@ -1,6 +1,6 @@
 #include "base/conv/tostring.h"
 
-#include "base/test/test.h"
+#include "../../util/test.h"
 
 using namespace test;
 using namespace Conv;
