+++ test/unit/base/gfx/color.cpp
@@ -1,6 +1,6 @@
 #include "base/gfx/color.h"
 
-#include "base/test/test.h"
+#include "../../util/test.h"
 
 using namespace test;
