+++ test/unit/base/refl/enum.cpp
@@ -1,6 +1,6 @@
 #include "base/refl/enum.h"
 
-#include "base/test/test.h"
+#include "../../util/test.h"
 
 using namespace test;
