+++ test/unit/base/refl/struct.cpp
@@ -1,6 +1,7 @@
 
 #include "base/refl/struct.h"
-#include "base/test/test.h"
+
+#include "../../util/test.h"
 
 using namespace test;
