+++ test/unit/base/style/parammerger.cpp
@@ -1,6 +1,6 @@
 #include "base/style/parammerger.h"
 
-#include "base/test/test.h"
+#include "../../util/test.h"
 
 #include "teststyle.h"
